pyproject.toml
[tool.poetry]
name = "spoural-excel"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "5.1.1"
djangorestframework = "3.15.2"
asgiref = "3.8.1"
flake8 = "7.1.1"
mccabe = "0.7.0"
mypy = "1.11.2"
mypy-extensions = "1.0.0"
psycopg2-binary = "2.9.9"
pycodestyle = "2.12.1"
pyflakes = "3.2.0"
sqlparse = "0.5.1"
typing-extensions = "4.12.2"
tzdata = "2024.2"
djangorestframework-simplejwt = "5.3.1"
pyjwt = "2.9.0"
django-cors-headers = "4.4.0"
openpyxl = "^3.1.5"
reportlab = "^4.2.5"
fpdf = "^1.7.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
